removing .class and created peckages
