+++ SumOfTheDigits/src/main/GetsNumbers.java
@@ -1,4 +1,4 @@
-package com.company;
+package main;
 
 public class GetsNumbers {
